# task2/gateway/cmd/gateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Gateway for the Collector service which can fetch information about GitHub repositories.",
        "title": "Gateway API",
        "contact": {},
        "license": {
            "url": "https://unlicense.org"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/repository/get": {
            "get": {
                "description": "get GitHub repository information by its owner and name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get repository by owner and name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Repository owner",
                        "name": "owner",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Repository name",
                        "name": "name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler_http.GetRepositoryResponse"
                        }
                    },
                    "400": {
                        "description": "Bad parameters",
                        "schema": {
                            "$ref": "#/definitions/internal_handler_http.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Repository not found",
                        "schema": {
                            "$ref": "#/definitions/internal_handler_http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler_http.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "internal_handler_http.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "internal_handler_http.GetRepositoryResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "forks_count": {
                    "type": "integer"
                },
                "license": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "stargazers_count": {
                    "type": "integer"
                }
            }
        }
    },
    "externalDocs": {
        "description": "OpenAPI Specification",
        "url": "https://swagger.io/resources/open-api/"
    }
}